backend/requirements.txt
Django>=4.2
psycopg2-binary
djangorestframework
django-filter
django-cors-headers
openai
